Implemented GetHeightAtPos function
GetHeightAtPos function gets world height of heightMap based on a
position in world space. This function can only be called on objects
that are guaranteed to be floating over the terrain, which is why the
AboveTerrain function was created.

The GetHeightAtPos converts the world pos of some otherPos to local
terrainCoordinates, then to uvCoords, samples the heightmap, then
converts that data to world coordinates once again.