Split Camera's UpdatePos innards into two separate functions
UpdatePos for Camera used to directly handled UserInput and Orbit state
code. Now, UpdatePos_UserInput handles UserInput pos update and
UpdatePos_Orbit handles Orbit pos update.

This was done to cleanup the code base in preparation for the new spring
mechanic that will be used with the camera's orbit state.